moving all utils to /agents/utils in either other.py or tableau.py
